tests/wpt/tests/css/css-ui/parsing/user-select-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 4: parsing user-select with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-4/#content-selection">
<meta name="assert" content="user-select supports only the grammar 'auto | text | none | contain | all'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("user-select", "visible");
test_invalid_value("user-select", "text contain");
</script>
</body>
</html>
